Move shared database logic to MySQL::DatabaseStatements
