+++ activerecord/lib/active_record/connection_adapters/trilogy_adapter.rb
@@ -182,70 +182,6 @@ module ActiveRecord
           connect
         end
 
-        def sync_timezone_changes(conn)
-          # Sync any changes since connection last established.
-          if default_timezone == :local
-            conn.query_flags |= ::Trilogy::QUERY_FLAGS_LOCAL_TIMEZONE
-          else
-            conn.query_flags &= ~::Trilogy::QUERY_FLAGS_LOCAL_TIMEZONE
-          end
-        end
-
-        def execute_batch(statements, name = nil)
-          statements = statements.map { |sql| transform_query(sql) }
-          combine_multi_statements(statements).each do |statement|
-            with_raw_connection do |conn|
-              raw_execute(statement, name)
-              conn.next_result while conn.more_results_exist?
-            end
-          end
-        end
-
-        def multi_statements_enabled?
-          !!@config[:multi_statement]
-        end
-
-        def with_multi_statements
-          if multi_statements_enabled?
-            return yield
-          end
-
-          with_raw_connection do |conn|
-            conn.set_server_option(::Trilogy::SET_SERVER_MULTI_STATEMENTS_ON)
-
-            yield
-          ensure
-            conn.set_server_option(::Trilogy::SET_SERVER_MULTI_STATEMENTS_OFF)
-          end
-        end
-
-        def combine_multi_statements(total_sql)
-          total_sql.each_with_object([]) do |sql, total_sql_chunks|
-            previous_packet = total_sql_chunks.last
-            if max_allowed_packet_reached?(sql, previous_packet)
-              total_sql_chunks << +sql
-            else
-              previous_packet << ";\n"
-              previous_packet << sql
-            end
-          end
-        end
-
-        def max_allowed_packet_reached?(current_packet, previous_packet)
-          if current_packet.bytesize > max_allowed_packet
-            raise ActiveRecordError,
-              "Fixtures set is too large #{current_packet.bytesize}. Consider increasing the max_allowed_packet variable."
-          elsif previous_packet.nil?
-            true
-          else
-            (current_packet.bytesize + previous_packet.bytesize + 2) > max_allowed_packet
-          end
-        end
-
-        def max_allowed_packet
-          @max_allowed_packet ||= show_variable("max_allowed_packet")
-        end
-
         def full_version
           schema_cache.database_version.full_version_string
         end
@@ -264,15 +200,6 @@ module ActiveRecord
 
         def default_prepared_statements
           false
-        end
-
-        def default_insert_value(column)
-          super unless column.auto_increment?
-        end
-
-        # https://mariadb.com/kb/en/analyze-statement/
-        def analyze_without_explain?
-          mariadb? && database_version >= "10.1.0"
         end
     end
   end
